+++ src/test/php/lang/ast/unittest/loader/Tests.php
@@ -1,0 +1,5 @@
+<?php namespace lang\ast\unittest\loader;
+
+class Tests {
+
+}